requirements.txt
# Core ML/DL
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Database
faiss-cpu>=1.7.4

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0

# Config & Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# LLM Integration
openai>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
